+++ testapps/GHIssues/src/main/java/ghissues/gh967/XyzService.java
@@ -1,0 +1,11 @@
+package ghissues.gh967;
+
+import act.util.SingletonBase;
+
+public class XyzService extends SingletonBase {
+
+    public String doService() {
+        return "xyz";
+    }
+
+}